devex-1197: fixing various paths after moving files
