+++ .circleci/config.yml
@@ -147,7 +147,7 @@ jobs:
           name: Admin User Setup
           command: |
             if [ "${CIRCLE_BRANCH}" != "prod" ] && [ "$DEPLOY_EMPTY_PERSISTENCE" != "true" ]; then
-              npx ts-node --transpile-only shared/admin-tools/user/setup-admin.ts
+              npx ts-node --transpile-only scripts/user/setup-admin.ts
             else
               echo "skipping…"
             fi
@@ -155,7 +155,7 @@ jobs:
           name: Test Users Setup
           command: |
             if [ "${CIRCLE_BRANCH}" != "prod" ] && [ "$DEPLOY_EMPTY_PERSISTENCE" != "true" ]; then
-              ./shared/admin-tools/user/setup-test-users.sh $ENV
+              ./scripts/user/setup-test-users.sh $ENV
             else
               echo "skipping…"
             fi
@@ -176,7 +176,7 @@ jobs:
                 --query 'DomainStatus.Endpoint' \
                 --output text)
               export ELASTICSEARCH_ENDPOINT="$ELASTICSEARCH_ENDPOINT"
-              npx ts-node --transpile-only ./shared/admin-tools/user/setup-glued-judges.ts
+              npx ts-node --transpile-only ./scripts/user/setup-glued-judges.ts
             else
               ./scripts/circleci/judge/bulk-import-judge-users.sh
             fi
@@ -740,11 +740,11 @@ jobs:
       - run:
           name: Test Users Setup
           command: |
-            DEPLOYING_COLOR="$CURRENT_COLOR" ./shared/admin-tools/user/setup-test-users.sh "$ENV"
+            DEPLOYING_COLOR="$CURRENT_COLOR" ./scripts/user/setup-test-users.sh "$ENV"
       - run:
           name: Judge Users Setup
           command: |
-            npx ts-node --transpile-only ./shared/admin-tools/user/setup-glued-judges.ts
+            npx ts-node --transpile-only ./scripts/user/setup-glued-judges.ts
       - run:
           name: Enable Dynamodb Streams
           command: |
