+++ docs/additional-resources/troubleshooting.md
@@ -252,7 +252,7 @@ Solution:
 - This solution should only be used on non-prod environments. Clear the dynamo tables and recreate the users and judges.
 
   ```bash
-  . ./shared/admin-tools/user/setup-test-users.sh "${ENV}"
+  . ./scripts/user/setup-test-users.sh "${ENV}"
   ```
 
   ```bash
