+++ scripts/run-once-scripts/restart-experimental-environment.sh
@@ -42,9 +42,9 @@ npm run deploy:api "${ENV}"
 npx ts-node --transpile-only ./web-api/reindex-dynamodb-records.js "${DESTINATION_TABLE}"
 
 # Setting up users
-npx ts-node --transpile-only shared/admin-tools/user/setup-admin.ts
+npx ts-node --transpile-only scripts/user/setup-admin.ts
 # shellcheck disable=SC1091
-. ./shared/admin-tools/user/setup-test-users.sh "${ENV}"
+. ./scripts/user/setup-test-users.sh "${ENV}"
 
 # Setting up Judge users
 ./scripts/circleci/judge/bulk-import-judge-users.sh
